+++ CHANGELOG.md
@@ -25,9 +25,15 @@
 - Added OpenTelemetry (OTEL) support for exporting traces, logs, and metrics over OTLP.
   - Added `--otel`, `--otel-logs`, and `--otel-service-name` global options.
   - Added metrics for tool installs, install steps, tool uninstalls, and plugin loads.
+- Improved the bin/shim registry layer, so that we can better detect executable name collisions.
 - Improved performance of checksum file hashing.
+- Updated `proto diagnose` to include warnings about executable name collisions.
 - **WASM API**
   - Added `DownloadPrebuiltOutput.post_script_args` field, which allows plugins to specify a list of arguments to pass to the post-install script when executing it.
+
+#### 🐞 Fixes
+
+- Fixed an issue where uninstalling a tool by version may leave around a stale/dangling bin file in `~/.proto/bin`.
 
 #### ⚙️ Internal